Use python types (int, bool,...) instead of generic numpy types (np.int, np.bool,...)

+++ tests/io_tests/test_hdf5.py
@@ -13,7 +13,7 @@ class TestHdf5:
         ('complex', {'key': 1.1j}, np.complex128(1.1j)),
         ('str', {'key': 'bla'}, 'bla'),
         ('none', {'key': None}, None),
-        ('np int', {'key': np.int(1)}, np.int32(1) if os.name == 'nt' else np.int64(1)),
+        ('np int', {'key': int(1)}, np.int32(1) if os.name == 'nt' else np.int64(1)),
         ('np float32', {'key': np.float32(1.1)}, np.float32(1.1)),
         ('np float64', {'key': np.float64(1.1)}, np.float64(1.1)),
         ('np complex64', {'key': np.complex64(1.1j)}, np.complex64(1.1j)),
